// src/shared/enums/resumeEnums.ts
export enum ResumeContactInfoEnum {
  portfolioLink = "portfolioLink",
  github = "github",
  linkedIn = "linkedIn",
  medium = "medium",
  twitter = "twitter",
}

export enum ResumeYearsOfExperience {
  internship = "internship",
  entryLevel = "entryLevel or junior",
  intermediate = "intermediate",
  firstLevelManagement = "first level management",
  middleLevelManagement = "middleLevelManagement",
  executive = "Executive or Senior",


// Executive or senior management.
// Middle management.
// First-level management.
// Intermediate or experienced (senior staff)
// Entry-level.
}

export enum TransformResumeFormType {
  dataToServer = "dataToServer",
  dataToClient = "dataToClient",
}

export enum ResumeAliases {
  header = "Header",
  fullName = "Full Name",
  jobTitle = "Job Title",
  avatar = "Avatar",
  contact = "Contact",
  emailAddress = "Email Address",
  location = "Location",
  phoneNumber = "Contact Number",
  summary = "Summary",
  yearsOfExperience = "Years Of Experience",
  experience = "Experience",
  skills = "Skills",
  education = "Education",
  certificate = "Certificate",
  awards = "Awards",
  publication = "Publication",
  references = "References",
  involvement = "Involvement",
  project = "Project",
}

export enum ActionTypeEnum {
  edit = "edit",
  add = "add",
  delete = "delete",
}

export enum SelectOrSelectedTemplate {
  select = "select",
  selected = "selected"
}